Add files via upload
I made a copy of the main file name anotherMain.rb, and made a few changes to run with the page2.erb file. now it covers the process of choosing a dialog, number, and time. then pops the stripe payment form, and finally redirects to charge.erb making the call in between the transition after the payment.